backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
httpx
